Remove Hashicorp support / TCP handling
I've decided to drop support for hashicorp/swim because my primary
reason for interfacing with its system was to test the behavior of
mine. There are better ways to test my implementation in addition
to the learning opportunities for writing robust tests.

SWIM does not *require* TCP as the paper is for UDP only. TCP is
used by Hashicorp for good reasons: increasing converagance rate
& efficacy, improves the suspicion mechanism, & circumvents bad
network configuration.